Rename libs -> lib & day.js -> dayjs.js
