+++ app/javascript/src/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import dayjs from "libs/day"; // eslint-disable-line
+import dayjs from "lib/dayjs"; // eslint-disable-line
 
 import { AuthProvider } from "contexts/auth";
 import { UserProvider } from "contexts/user";
