+++ config/webpack/alias.js
@@ -6,7 +6,7 @@ module.exports = {
       components: "src/components",
       constants: "src/constants",
       contexts: "src/contexts",
-      libs: "src/libs",
+      lib: "src/lib",
       reducers: "src/reducers",
       neetoui: "@bigbinary/neetoui",
       neetoicons: "@bigbinary/neeto-icons",
